+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/mixin/EntityAccessor.java
@@ -1,0 +1,13 @@
+package com.github.khanshoaib3.minecraft_access.mixin;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.text.Text;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+
+@Mixin(Entity.class)
+public interface EntityAccessor {
+    @Invoker
+    Text callGetName();
+}